Base Django project
This adds a base django project that serves a simple page frontend page
with the Laika Fleet Systems logo. The backend is intended to be a simple
API that can be used by the backend to grab data; the frontend will do
most of the heavy lifting of the app.

Instructions are provided on how to do a quickstart developer setup.
NOTE: This setup may change in the future.